Broke up App.js functions and html return into components
