+++ src/components/Footer/Footer.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export function Footer() {
+    return (
+        <footer>
+            <p> Made with ❤️ by Abdul-Kadir Coskun</p>
+        </footer>
+    );
+}
