+++ src/components/Header/Header.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export function Header() {
+    return (
+        <header>
+            <h1>Gif Converter</h1>
+        </header>
+    );
+}
